creating committee and "my legislator" emails extracted to separate files for manageability
